requirements.txt
annotated-types==0.6.0
ansible==9.1.0
ansible-core==2.16.2
ansible-runner==2.3.4
anyio==4.2.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
docutils==0.20.1
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.3
packaging==23.2
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-daemon==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
resolvelib==1.0.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
